Move run in terminal tool telemetry into separate file
